+++ aiohomematic/interfaces/model.py
@@ -1809,7 +1809,7 @@ class DeviceWeekProfileProtocol(Protocol):
 
     @property
     @abstractmethod
-    def week_profile(self) -> WeekProfileProtocol[dict[Any, Any]] | None:
+    def week_profile(self) -> WeekProfileProtocol[Any] | None:
         """Return the week profile."""
 
     @abstractmethod
@@ -2091,7 +2091,7 @@ class HubProtocol(Protocol):
 
 
 @runtime_checkable
-class WeekProfileProtocol[SCHEDULE_DICT_T: dict[Any, Any]](Protocol):
+class WeekProfileProtocol[SCHEDULE_DICT_T](Protocol):
     """
     Protocol for week profile operations.